requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
anthropic>=0.30.0
groq>=0.9.0
pypdf>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
